fix(lint): keep stdout retry helper isolated
