+++ crates/vize/src/commands/lint/mod.rs
@@ -3,6 +3,7 @@
 mod args;
 mod collect;
 mod cross_file;
+mod stdout;
 
 #[cfg(test)]
 mod tests;
@@ -14,19 +15,16 @@ use collect::{collect_lint_files, is_standalone_html_path, resolve_lint_config_p
 use cross_file::{build_cross_file_lint_output, merge_lint_result};
 use rayon::prelude::*;
 use std::fs;
-use std::io::{self, Write};
+use std::io::Write;
 use std::path::{Path, PathBuf};
 use std::sync::Mutex;
 use std::sync::atomic::{AtomicUsize, Ordering};
-use std::time::Duration;
-use std::time::Instant;
+use std::time::{Duration, Instant};
 use vize_carton::{String, ToCompactString, cstr, profile, profiler::global_profiler};
 use vize_curator::profile::{
     ProfileFileRow, ProfilePhase, ProfilePhaseKind, ProfileReport, print_profile_report,
 };
-use vize_patina::{
-    HelpLevel, JsxLang, LintPreset, Linter, OutputFormat, format_results, format_summary,
-};
+use vize_patina::{HelpLevel, JsxLang, LintPreset, Linter, OutputFormat, format_results};
 
 pub fn run(args: LintArgs) {
     let start = Instant::now();
@@ -249,7 +247,7 @@ pub fn run(args: LintArgs) {
             format_results(&lint_results, &sources, format)
         );
         if !output.trim().is_empty() {
-            write_stdout(output.as_bytes());
+            stdout::write(output.as_bytes());
         }
     }
     let output_time = output_start.elapsed();
@@ -267,19 +265,15 @@ pub fn run(args: LintArgs) {
     // Print summary
     let elapsed = start.elapsed();
     if format == OutputFormat::Text {
-        write_stdout(
-            cstr!(
-                "\n{}\n",
-                format_summary(total_errors, total_warnings, files.len())
-            )
-            .as_bytes(),
+        stdout::write_text_summary(
+            total_errors,
+            total_warnings,
+            files.len(),
+            elapsed,
+            args.cross_file_tree
+                .then_some(cross_file_tree.as_deref())
+                .flatten(),
         );
-        write_stdout(cstr!("Linted {} files in {:.4?}\n", files.len(), elapsed).as_bytes());
-        if args.cross_file_tree
-            && let Some(tree) = cross_file_tree.as_deref()
-        {
-            write_stdout(cstr!("\n{tree}\n").as_bytes());
-        }
     }
 
     if args.profile {
@@ -401,59 +395,6 @@ pub fn run(args: LintArgs) {
         eprintln!("\nToo many warnings ({} > max {})", total_warnings, max);
         std::process::exit(1);
     }
-}
-
-fn write_stdout(bytes: &[u8]) {
-    let mut stdout = io::stdout().lock();
-    if let Err(error) = write_all_retry(&mut stdout, bytes) {
-        handle_stdout_error(error);
-    }
-}
-
-fn write_all_retry<W: Write>(writer: &mut W, bytes: &[u8]) -> io::Result<()> {
-    const MAX_CONSECUTIVE_WOULD_BLOCK: usize = 1024;
-
-    let mut written = 0;
-    let mut would_block_count = 0;
-    while written < bytes.len() {
-        match writer.write(&bytes[written..]) {
-            Ok(0) => {
-                return Err(io::Error::new(
-                    io::ErrorKind::WriteZero,
-                    "failed to write lint output",
-                ));
-            }
-            Ok(count) => {
-                written += count;
-                would_block_count = 0;
-            }
-            Err(error)
-                if matches!(
-                    error.kind(),
-                    io::ErrorKind::Interrupted | io::ErrorKind::WouldBlock
-                ) =>
-            {
-                would_block_count += usize::from(error.kind() == io::ErrorKind::WouldBlock);
-                if would_block_count > MAX_CONSECUTIVE_WOULD_BLOCK {
-                    return Err(error);
-                }
-                std::thread::sleep(Duration::from_millis(1));
-            }
-            Err(error) => return Err(error),
-        }
-    }
-    Ok(())
-}
-
-fn handle_stdout_error(error: io::Error) -> ! {
-    if error.kind() == io::ErrorKind::BrokenPipe {
-        std::process::exit(0);
-    }
-    eprintln!(
-        "\x1b[31mError:\x1b[0m failed to write lint output: {}",
-        error
-    );
-    std::process::exit(1);
 }
 
 fn jsx_lang_for_path(path: &Path) -> Option<JsxLang> {
